events: move all layer reordering code to events store
