+++ src/stores/Store.js
@@ -310,8 +310,9 @@ class Store {
         this.layerVisibleMap.set(layer.id, isVisible);
     }
 
-    // monkey patch this to execute some logic after layers have been loaded
-    doAfterLayersLoaded = () => { };
+    _doAfterLayersLoaded() { 
+        // overwrite this to execute some logic after layers have been loaded
+     };
 
     async load(mapViewDiv) {
         this.appState.loadingMessage('Loading map.');
@@ -352,7 +353,7 @@ class Store {
             .then(_ => {
                 this.loadFilters();
                 this.loadCharts();
-                this.doAfterLayersLoaded();
+                this._doAfterLayersLoaded(this.view);
             })
             .catch(er => {
                 this.appState.onError(er, 'Could not load layers, do you have access to the data?')
@@ -515,7 +516,7 @@ decorate(Store, {
     _onMouseLeave: action.bound,
     startAutoplayBookmarks: action.bound,
     stopAutoplayBookmarks: action.bound,
-    doAfterLayersLoaded: action.bound,
+    _doAfterLayersLoaded: action.bound,
 });
 
 export default Store;
